requirements.txt
python-decouple==3.3
dj-database-url==0.5.0
psycopg2==2.8.4
drf-yasg==1.17.1
Django==3.0.5
entrypoints==0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.2.0
mixer==6.1.3
asgiref==3.2.7
astroid==2.3.3
coreapi==2.3.3
coreschema==0.0.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
